Adiciona serviço de cep na camada de infraestrutura
