--- src/Backend/SistemaDeCadastro.Application/Servicos/CepServices/ICepServices.cs
@@ -1,8 +1,0 @@
-namespace SistemaDeCadastro.Application.Servicos.CepServices;
-
-    public interface ICepServices
-    {
-        bool ValidarCep(string cep);
-
-        Task<EnderecoJson> RecuperarEndereco(string cep);
-    }
