+++ src/Backend/SistemaDeCadastro.Domain/Servicos/ViaCep/EnderecoJson.cs
@@ -1,0 +1,13 @@
+namespace SistemaDeCadastro.Domain.Servicos.ViaCep;
+
+public record EnderecoJson(
+    [property: JsonPropertyName("cep")] string Cep,
+    [property: JsonPropertyName("logradouro")] string Logradouro,
+    [property: JsonPropertyName("complemento")] string Complemento,
+    [property: JsonPropertyName("bairro")] string Bairro,
+    [property: JsonPropertyName("localidade")] string Localidade,
+    [property: JsonPropertyName("uf")] string Uf,
+    [property: JsonPropertyName("ibge")] string Ibge,
+    [property: JsonPropertyName("gia")] string Gia,
+    [property: JsonPropertyName("ddd")] string Ddd,
+    [property: JsonPropertyName("siafi")] string Siafi);
