+++ src/Backend/SistemaDeCadastro.Domain/Servicos/ViaCep/IViaCep.cs
@@ -1,0 +1,8 @@
+namespace SistemaDeCadastro.Domain.Servicos.ViaCep;
+
+public interface IViaCep
+{
+    bool ValidarCep(string cep);
+
+    Task<EnderecoJson> RecuperarEndereco(string cep);
+}
