+++ src/Backend/SistemaDeCadastro.Domain/Usings.cs
@@ -1,6 +1,9 @@
+global using System.Text.Json.Serialization;
+global using Microsoft.Extensions.Configuration;
 global using SistemaDeCadastro.Domain.Entidades.Base;
 global using SistemaDeCadastro.Domain.ValueObjects;
 global using SistemaDeCadastro.Domain.ValueObjects.Base;
 global using SistemaDeCadastro.Domain.Enum;
-global using Microsoft.Extensions.Configuration;
+
+
