+++ src/Backend/SistemaDeCadastro.Infrastructure/Usings.cs
@@ -1,5 +1,6 @@
 global using System.Reflection;
 global using System.Data;
+global using System.Text.Json;
 global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.AspNetCore.Builder;
@@ -13,10 +14,14 @@ global using SistemaDeCadastro.Domain.ValueObjects;
 global using SistemaDeCadastro.Domain.Repositorios;
 global using SistemaDeCadastro.Domain.Repositorios.Cadastro;
 global using SistemaDeCadastro.Domain.Repositorios.Pessoa;
+global using SistemaDeCadastro.Domain.Servicos.ViaCep;
 
 global using SistemaDeCadastro.Infrastructure.AcessoRepositorio;
 global using SistemaDeCadastro.Infrastructure.AcessoRepositorio.Repositorio;
+global using SistemaDeCadastro.Infrastructure.Servicos.ViaCep;
 
+global using SistemaDeCadastro.Exceptions;
+global using SistemaDeCadastro.Exceptions.ExceptionsBase;
